+++ R/utils.R
@@ -31,10 +31,15 @@
 #' @examples
 #' .emr_expr_vars("a + b")
 #'
-#' @noRd
+#' @keywords internal
 #' @export
 .emr_expr_vars <- function(expr) {
-    all.vars(as.list(parse(text = expr))[[1]])
+    res <- c()
+
+    if (!is.null(expr) && expr != "") {
+        res <- all.vars(as.list(parse(text = expr))[[1]])
+    }
+    return(res)
 }
 
 .emr_getOption <- function(x, default = NULL) {