+++ R/zzz.R
@@ -1,8 +1,7 @@
 #' An environment for storing the package global variables
 #'
-#'
+#' @keywords internal
 #' @export
-#' @noRd
 .naryn <- new.env(parent = emptyenv())
 .naryn$entries <- list()
 .naryn$entries_timestamp <- list()